+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Density returns 0 instead of NAN outside (0,1)
 - Changed first argument name of dlogitnorm from q to x
+- Document argument n of rlogitnorm
 
 # logitnorm 0.8.36
 Remove the library call to MASS.